Renames ai_routing to ai_api_routing
