+++ qiskit_ibm_transpiler/wrappers/__init__.py
@@ -32,7 +32,7 @@ Classes
    TranspileAPI
 """
 
-from .ai_routing import AIRoutingAPI
+from .ai_api_routing import AIRoutingAPI
 from .ai_local_routing import AILocalRouting
 from .ai_api_synthesis import AICliffordAPI, AILinearFunctionAPI, AIPermutationAPI
 from .base import BackendTaskError, QiskitTranspilerService
